+++ mitmproxy/certs.py
@@ -32,6 +32,7 @@ logger = logging.getLogger(__name__)
 # Default expiry must not be too long: https://github.com/mitmproxy/mitmproxy/issues/815
 CA_EXPIRY = datetime.timedelta(days=10 * 365)
 CERT_EXPIRY = datetime.timedelta(days=365)
+CRL_EXPIRY = datetime.timedelta(days=7)
 
 # Generated with "openssl dhparam". It's too slow to generate this on startup.
 DEFAULT_DHPARAM = b"""
@@ -189,6 +190,22 @@ class Cert(serializable.Serializable):
         else:
             return x509.GeneralNames(sans)
 
+    @property
+    def crl_distribution_points(self) -> list[str]:
+        try:
+            ext = self._cert.extensions.get_extension_for_class(
+                x509.CRLDistributionPoints
+            ).value
+        except x509.ExtensionNotFound:
+            return []
+        else:
+            return [
+                dist_point.full_name[0].value
+                for dist_point in ext
+                if dist_point.full_name
+                and isinstance(dist_point.full_name[0], x509.UniformResourceIdentifier)
+            ]
+
 
 def _name_to_keyval(name: x509.Name) -> list[tuple[str, str]]:
     parts = []
@@ -287,6 +304,7 @@ def dummy_cert(
     commonname: str | None,
     sans: Iterable[x509.GeneralName],
     organization: str | None = None,
+    crl_url: str | None = None,
 ) -> Cert:
     """
     Generates a dummy certificate.
@@ -296,6 +314,7 @@ def dummy_cert(
     commonname: Common name for the generated certificate.
     sans: A list of Subject Alternate Names.
     organization: Organization name for the generated certificate.
+    crl_url: URL of CRL distribution point
 
     Returns cert if operation succeeded, None if not.
     """
@@ -337,8 +356,47 @@ def dummy_cert(
     # https://datatracker.ietf.org/doc/html/rfc5280#section-4.2.1.2 states
     # that SKI is optional for the leaf cert, so we skip that.
 
+    if crl_url:
+        builder = builder.add_extension(
+            x509.CRLDistributionPoints(
+                [
+                    x509.DistributionPoint(
+                        [x509.UniformResourceIdentifier(crl_url)],
+                        relative_name=None,
+                        crl_issuer=None,
+                        reasons=None,
+                    )
+                ]
+            ),
+            critical=False,
+        )
+
     cert = builder.sign(private_key=privkey, algorithm=hashes.SHA256())  # type: ignore
     return Cert(cert)
+
+
+def dummy_crl(
+    privkey: rsa.RSAPrivateKey,
+    cacert: x509.Certificate,
+) -> bytes:
+    """
+    Generates an empty CRL signed with the CA key
+
+    privkey: CA private key
+    cacert: CA certificate
+
+    Returns a CRL DER encoded
+    """
+    builder = x509.CertificateRevocationListBuilder()
+    builder = builder.issuer_name(cacert.issuer)
+
+    now = datetime.datetime.now()
+    builder = builder.last_update(now - datetime.timedelta(days=2))
+    builder = builder.next_update(now + CRL_EXPIRY)
+
+    builder = builder.add_extension(x509.CRLNumber(1000), False)  # meaningless number
+    crl = builder.sign(private_key=privkey, algorithm=hashes.SHA256())
+    return crl.public_bytes(serialization.Encoding.DER)
 
 
 @dataclass(frozen=True)
@@ -375,11 +433,13 @@ class CertStore:
         default_privatekey: rsa.RSAPrivateKey,
         default_ca: Cert,
         default_chain_file: Path | None,
+        default_crl: bytes,
         dhparams: DHParams,
     ):
         self.default_privatekey = default_privatekey
         self.default_ca = default_ca
         self.default_chain_file = default_chain_file
+        self.default_crl = default_crl
         self.default_chain_certs = (
             [
                 Cert(c)
@@ -448,11 +508,12 @@ class CertStore:
         dh = cls.load_dhparam(dhparam_file)
         certs = x509.load_pem_x509_certificates(raw)
         ca = Cert(certs[0])
+        crl = dummy_crl(key, ca._cert)
         if len(certs) > 1:
             chain_file: Path | None = ca_file
         else:
             chain_file = None
-        return cls(key, ca, chain_file, dh)
+        return cls(key, ca, chain_file, crl, dh)
 
     @staticmethod
     @contextlib.contextmanager
@@ -593,6 +654,7 @@ class CertStore:
         commonname: str | None,
         sans: Iterable[x509.GeneralName],
         organization: str | None = None,
+        crl_url: str | None = None,
     ) -> CertStoreEntry:
         """
         commonname: Common name for the generated certificate. Must be a
@@ -601,6 +663,8 @@ class CertStore:
         sans: A list of Subject Alternate Names.
 
         organization: Organization name for the generated certificate.
+
+        crl_url: URL of CRL distribution point
         """
         sans = _fix_legacy_sans(sans)
 
@@ -623,6 +687,7 @@ class CertStore:
                     commonname,
                     sans,
                     organization,
+                    crl_url,
                 ),
                 privatekey=self.default_privatekey,
                 chain_file=self.default_chain_file,